+++ api/applib/src/main/java/org/apache/isis/applib/services/clock/VirtualClock_withOffset.java
@@ -16,22 +16,29 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.isis.applib.clock;
+package org.apache.isis.applib.services.clock;
 
 import java.time.Instant;
+import java.util.Objects;
 
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-final class VirtualClock_system implements VirtualClock {
+final class VirtualClock_withOffset implements VirtualClock {
 
-    private static final long serialVersionUID = 4019699893189814294L;
-
+    private static final long serialVersionUID = -2589204298085221985L;
+    
+    /**
+     * Amount of time (milli seconds) this clock is offset into the future 
+     * with respect to the actual (system) time.
+     */
+    private final long millisOffset;
+    
     @Override
     public Instant now() {
-        return Instant.now();
+        return Instant.now().plusMillis(millisOffset);
     }
-
+    
     // -- TO STRING, EQUALS, HASHCODE
     
     @Override
@@ -44,14 +51,20 @@ final class VirtualClock_system implements VirtualClock {
         if(obj==null) {
             return false;
         }
-        // equal if same class
-        return this.getClass().equals(obj.getClass());
+        // equal if same class and same millisOffset
+        if(!Objects.equals(this.getClass(), obj.getClass())) {
+            return false;
+        }
+        if(!Objects.equals(this.millisOffset, ((VirtualClock_withOffset)obj).millisOffset)) {
+            return false;
+        }
+        return true;
     }
     
     @Override
     public int hashCode() {
-        // equal if same class
-        return this.getClass().hashCode();
+        // equal if same class and same millisOffset
+        return Objects.hash(this.getClass(), millisOffset);
     }
 
 }
