+++ testing/fixtures/applib/src/main/java/org/apache/isis/testing/fixtures/applib/IsisIntegrationTestAbstractWithFixtures.java
@@ -26,9 +26,9 @@ import org.junit.jupiter.api.BeforeEach;
 
 import org.apache.isis.applib.services.clock.ClockService;
 import org.apache.isis.testing.fixtures.applib.api.PersonaWithBuilderScript;
-import org.apache.isis.testing.fixtures.applib.clock.Clock;
-import org.apache.isis.testing.fixtures.applib.clock.FixtureClock;
-import org.apache.isis.testing.fixtures.applib.clock.TickingFixtureClock;
+import org.apache.isis.testing.fixtures.applib.services.clock.Clock;
+import org.apache.isis.testing.fixtures.applib.services.clock.FixtureClock;
+import org.apache.isis.testing.fixtures.applib.services.clock.TickingFixtureClock;
 import org.apache.isis.testing.fixtures.applib.fixturescripts.BuilderScriptAbstract;
 import org.apache.isis.testing.fixtures.applib.fixturescripts.FixtureScript;
 import org.apache.isis.testing.fixtures.applib.fixturescripts.FixtureScripts;
