+++ testing/fixtures/applib/src/main/java/org/apache/isis/testing/fixtures/applib/clock/FixtureClock.java
@@ -16,7 +16,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.isis.testing.fixtures.applib.clock;
+package org.apache.isis.testing.fixtures.applib.services.clock;
 
 import java.time.Instant;
 import java.util.Calendar;
