+++ testing/fixtures/applib/src/main/java/org/apache/isis/testing/fixtures/applib/clock/TickingFixtureClock.java
@@ -16,7 +16,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.isis.testing.fixtures.applib.clock;
+package org.apache.isis.testing.fixtures.applib.services.clock;
 
 import java.sql.Timestamp;
 import java.time.Instant;
